Some fixes for OS X: - Find enet properly - Type conversion of size_t's in JS
git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@6324 3db68df2-c116-0410-a063-a993310a9797

+++ source/scripting/SynchedJSObject.cpp
@@ -5,13 +5,13 @@
 #include "ScriptCustomTypes.h"
 
 template <>
-CStrW ToNetString(const unsigned &val)
+CStrW ToNetString(const size_t &val)
 {
 	return CStrW(val);
 }
 
 template <>
-void SetFromNetString(unsigned &val, const CStrW& string)
+void SetFromNetString(size_t &val, const CStrW& string)
 {
 	val=string.ToUInt();
 }
